feat: transition plot single point
This is an alternative to the existing full stateless transition plot.
The following implementation uses transition_data.txt for the transition
plot and andie_data.txt for the next temperature trace

- The dashboard service adds the record received along with an id to the
corresponding file (transition_data.txt for get_transition_data_single,
and andie.txt for get_next_temperature)
- The dashboard implements an id check by extracting the last record
received from the file, if the id is not the current id, the plot is
updated.
- For the next temperature, we update after updating transition to look
for the newest prediction to be plotted as the ANDiE next temperature
trace in the transition plot.